+++ libraries/exoplayer_ima/src/main/java/androidx/media3/exoplayer/ima/AdTagLoader.java
@@ -17,11 +17,11 @@ package androidx.media3.exoplayer.ima;
 
 import static androidx.media3.common.Player.COMMAND_GET_VOLUME;
 import static androidx.media3.common.util.Util.msToUs;
-import static com.google.common.base.Preconditions.checkArgument;
 import static androidx.media3.exoplayer.ima.ImaUtil.BITRATE_UNSET;
 import static androidx.media3.exoplayer.ima.ImaUtil.TIMEOUT_UNSET;
 import static androidx.media3.exoplayer.ima.ImaUtil.getAdGroupTimesUsForCuePoints;
 import static androidx.media3.exoplayer.ima.ImaUtil.getImaLooper;
+import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.Math.max;
@@ -494,8 +494,9 @@ import java.util.Objects;
     int windowIndex = player.getCurrentMediaItemIndex();
     Timeline.Window window = new Timeline.Window();
     timeline.getWindow(windowIndex, window);
-    checkArgument(window.firstPeriodIndex == window.lastPeriodIndex);
-    // may be TIME_UNSET for placeholder
+    checkArgument(
+        window.firstPeriodIndex == window.lastPeriodIndex,
+        "Only single-period Timelines are supported.");
     long contentDurationUs = timeline.getPeriod(window.firstPeriodIndex, period).durationUs;
 
     contentDurationMs = Util.usToMs(contentDurationUs);